+++ web/src/beta/components/fields/CameraField/CapturePanel/index.tsx
@@ -7,7 +7,7 @@ import { Camera } from "@reearth/beta/utils/value";
 import { useT } from "@reearth/services/i18n";
 import { styled } from "@reearth/services/theme";
 
-import PanelCommon from "../PanelCommon";
+import PanelCommon from "../../common/PanelCommon";
 import type { RowType } from "../types";
 
 type Props = {
