+++ web/src/beta/components/fields/DateTimeField/EditPanel/hooks.ts
@@ -1,5 +1,7 @@
 import moment from "moment-timezone";
-import { useCallback, useState } from "react";
+import { useCallback, useMemo, useState } from "react";
+
+import { getUniqueTimezones } from "@reearth/beta/utils/moment-timezone";
 
 type Props = {
   onChange?: (value?: string | undefined) => void;
@@ -13,7 +15,6 @@ type TimezoneInfo = {
 export default ({ onChange }: Props) => {
   const [time, setTime] = useState<string>("");
   const [date, setDate] = useState<string>("");
-  const [timezones] = useState(moment.tz.names());
   const [selectedTimezone, setSelectedTimezone] = useState("+0:00");
 
   const handleTimeChange = useCallback((newValue: string | undefined) => {
@@ -26,34 +27,9 @@ export default ({ onChange }: Props) => {
     setDate(newValue);
   }, []);
 
-  const getUniqueTimezones = useCallback((timezones: string[]): TimezoneInfo[] => {
-    const uniqueTimezones: TimezoneInfo[] = [];
-    const seenOffsets = new Set<string>();
-    timezones.forEach(timezone => {
-      const offset = moment.tz(timezone).utcOffset() / 60;
-
-      if (Number.isInteger(offset)) {
-        const offsetString = offset >= 0 ? `+${offset}` : `${offset}`;
-        const offsetTimezone = `${offsetString}:00`;
-
-        if (!seenOffsets.has(offsetTimezone)) {
-          uniqueTimezones.push({ timezone, offset: offsetTimezone });
-          seenOffsets.add(offsetTimezone);
-        }
-      }
-    });
-
-    uniqueTimezones.sort((a, b) => {
-      const offsetA = parseInt(a.offset);
-      const offsetB = parseInt(b.offset);
-      return offsetA - offsetB;
-    });
-
-    return uniqueTimezones;
+  const offsetFromUTC: TimezoneInfo[] = useMemo(() => {
+    return getUniqueTimezones(moment.tz.names());
   }, []);
-
-  const zones: string[] = moment.tz.names();
-  const offsetFromUTC: TimezoneInfo[] = getUniqueTimezones(zones);
 
   const handleApplyChange = useCallback(() => {
     const selectedTimezoneInfo = offsetFromUTC.find(info => info.timezone === selectedTimezone);
@@ -64,11 +40,10 @@ export default ({ onChange }: Props) => {
   }, [date, time, selectedTimezone, onChange, offsetFromUTC]);
 
   return {
-    timezones,
     date,
     time,
     selectedTimezone,
-    getUniqueTimezones,
+    offsetFromUTC,
     onTimeChange: handleTimeChange,
     onTimezoneSelect: setSelectedTimezone,
     onDateChange: handleDateChange,
