+++ web/src/beta/components/fields/common/PanelCommon.tsx
